fastapi/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-core>=0.1.0
langchain-deepseek>=0.0.1
asyncio>=3.4.3
python-multipart>=0.0.6
alembic>=1.12.0
httpx>=0.24.0